% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMultipleCohortSubsetDefinitions.R
\name{addMultipleCohortSubsetDefinitions}
\alias{addMultipleCohortSubsetDefinitions}
\title{Add Multiple Cohort Subset Definitions with Progress Bar}
\usage{
addMultipleCohortSubsetDefinitions(
  cohortDefinitionSet,
  targetCohortIds,
  prefix,
  ...
)
}
\arguments{
\item{cohortDefinitionSet}{The initial cohort definition set to which subset definitions will be added.}

\item{targetCohortIds}{A vector of target cohort IDs that the subset definitions will be applied to.}

\item{prefix}{A string prefix for the names of the subset definitions.}

\item{...}{Additional arguments representing sequences of subset operators.}
}
\value{
The modified cohort definition set with added subset definitions.
}
\description{
This function iteratively adds cohort subset definitions to a given cohort definition set.
It processes multiple sequences of subset operators, first adding each operator from the first sequence individually,
and then combining each operator from the first sequence with each operator from the subsequent sequences in turn.
A progress bar and console messages indicate the progress of the operation.
}
\examples{
# Usage:
# cohortDefinitionSetWithSubset <- addMultipleCohortSubsetDefinitions(
#   cohortDefinitionSet,
#   targetOrIndicationCohortIds,
#   "CustomPrefix",
#   subsetOperatorSequence1,
#   subsetOperatorSequence2
#   # ... Add more sequences if necessary
# )
}
