% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignParentId.R
\name{assignParentId}
\alias{assignParentId}
\title{Assign Parent ID Based on Name Substring Matching}
\usage{
assignParentId(df)
}
\arguments{
\item{df}{A data frame with at least two columns: \code{id} and \code{name}.}
}
\value{
A data frame with an additional column \code{parentId} containing the assigned parent IDs.
}
\description{
This function assigns \code{parentId} for each entry in a data frame based on the logic that the
parent ID should be the ID of the entry that has the name which is a substring of the current
name, is shorter than the current name, and has the most characters matching contiguously
with the current name.
}
