% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcatenteFeatureCohortSummaryToSubject.R
\name{concatenteFeatureCohortSummaryToSubject}
\alias{concatenteFeatureCohortSummaryToSubject}
\title{Concatenate Feature Cohort Summary to Subject Table}
\usage{
concatenteFeatureCohortSummaryToSubject(
  connection,
  subjectTableName,
  subjectTableDatabaseSchema = NULL,
  subjectTableIsTemp = FALSE,
  subjectStartDate = NULL,
  subjectEndDate = NULL,
  featureCohortTableName,
  featureCohortDatabaseSchema,
  featureCohortIsTemp = FALSE,
  featureCohortId,
  tempEmulationSchema = NULL,
  prefix = NULL
)
}
\arguments{
\item{connection}{Database connection object.}

\item{subjectTableName}{Name of the subject table.}

\item{subjectTableDatabaseSchema}{(Optional) Database schema that contains the wide subject table.}

\item{subjectTableIsTemp}{Boolean flag indicating if the subject table is temporary.}

\item{subjectStartDate}{(Optional) Date field in the subject table to left-censor the occurrence check.}

\item{subjectEndDate}{(Optional) Date field in the subject table to right-censor the occurrence check.}

\item{featureCohortTableName}{Name of the feature cohort table.}

\item{featureCohortDatabaseSchema}{Database schema containing the feature cohort table.}

\item{featureCohortIsTemp}{Boolean flag indicating if the feature cohort table is temporary.}

\item{featureCohortId}{Identifier for the feature cohort.}

\item{tempEmulationSchema}{(Optional) Schema for temporary emulation.}

\item{prefix}{(Optional) Prefix for the new columns added to the subject table. If NULL, then prefix will be the feature cohort id such as c1234}
}
\value{
A tibble listing the new columns added to the subject table.
}
\description{
This function performs a left join on a subject table with a feature cohort table
and adds new columns to the subject table based on the feature cohorts.
The subject table must have one row per subject and be located in a remote database,
and must contain a 'subject_id' column at the minimum.
}
\details{
If 'subjectStartDate' and 'subjectEndDate' are provided, additional columns are added
to the subject table to reflect these date limits; otherwise, only the occurrence dates
of the feature cohort are returned without any start or end date limits.
}
