% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCombinations.R
\name{createCombinations}
\alias{createCombinations}
\title{Generate Combinations of Array Elements}
\usage{
createCombinations(array)
}
\arguments{
\item{array}{A character vector containing unique elements for which combinations are to be generated.}
}
\value{
A tibble with two columns: \code{array} containing the individual elements, and \code{combinatrix} containing the combinations.
}
\description{
This function generates all possible combinations of the elements in a given array.
For each combination, it creates a tibble row for every element in the array that is part of that combination.
The resulting tibble contains two columns: one for the individual elements and another for the combinations they appear in.
}
