% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNonTemporalCovariateSettings.R
\name{createNonTemporalCovariateSettings}
\alias{createNonTemporalCovariateSettings}
\title{Create the non temporal covariate settings}
\usage{
createNonTemporalCovariateSettings(
  excludedCovariateConceptIds = c(),
  includedCovariateIds = c(),
  includedCovariateConceptIds = c(),
  addDescendantsToExclude = FALSE,
  startDayWindow1 = 0,
  endDayWindow1 = 9999,
  startDayWindow2 = NULL,
  endDayWindow2 = NULL,
  startDayWindow3 = NULL,
  endDayWindow3 = NULL,
  useDemographicsGender = TRUE,
  useDemographicsAgeGroup = TRUE,
  useDemographicsRace = TRUE,
  useDemographicsEthnicity = TRUE,
  useConditionGroupEra = TRUE,
  useDrugGroupEra = TRUE,
  useProcedureOccurrence = TRUE,
  useDeviceExposureLongTerm = TRUE,
  useMeasurement = TRUE,
  useMeasurementValue = TRUE,
  useMeasurementRangeGroup = TRUE,
  useObservation = TRUE,
  useDistinctConditionCount = TRUE,
  useDistinctIngredientCount = TRUE,
  useDistinctProcedureCount = TRUE,
  useDistinctMeasurementCount = TRUE,
  useVisitCount = TRUE,
  useVisitConceptCount = TRUE
)
}
\arguments{
\item{excludedCovariateConceptIds}{A list of conceptIds to exclude from featureExtraction.  These
should include all concept_ids that were used to define the
xSpec model (default=NULL)}

\item{includedCovariateIds}{A list of covariate IDs that should be restricted to.}

\item{includedCovariateConceptIds}{A list of covariate concept IDs that should be restricted to.}

\item{addDescendantsToExclude}{Should descendants of excluded concepts also be excluded?
(default=FALSE)}

\item{startDayWindow1}{The day to start time window 1 for feature extraction}

\item{endDayWindow1}{The day to end time window 1 for feature extraction}

\item{startDayWindow2}{The day to start time window 2 for feature extraction}

\item{endDayWindow2}{The day to end time window 2 for feature extraction}

\item{startDayWindow3}{The day to start time window 3 for feature extraction}

\item{endDayWindow3}{The day to end time window 3 for feature extraction #'}
}
\description{
Create the non temporal covariate settings using three time windows
}
\details{
Create the non temporal covariate settings using three time windows. It uses FeatureExtraction::createCovariateSettings
}
