% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditCohortGeneratorIncrementalLog.R
\name{editCohortGeneratorIncrementalLog}
\alias{editCohortGeneratorIncrementalLog}
\title{Edit Incremental Cohort Generator Log}
\usage{
editCohortGeneratorIncrementalLog(
  incrementalFolder,
  incrementalFileName = "GeneratedCohorts.csv",
  deleteCohortIds = NULL,
  keepCohortIds = NULL
)
}
\arguments{
\item{incrementalFolder}{A string specifying the path to the folder where the
incremental log file is located.}

\item{incrementalFileName}{A string specifying the name of the CSV file to be
edited. Defaults to "GeneratedCohorts.csv".}

\item{deleteCohortIds}{A numeric vector of cohort IDs to be removed from the
log file. If NULL, no deletions are made. Defaults to NULL.}
}
\value{
The function does not return anything, but it updates the specified
CSV file in the given folder.
}
\description{
This function edits a CSV file containing information about cohorts,
typically used in a cohort generation process. It allows for incremental
updates to the log file by either adding new data or deleting specified cohorts.
}
\examples{
editCohortGeneratorIncrementalLog("path/to/folder", deleteCohortIds = c(101, 102))

}
