% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteCohortCovariateCharacterization.R
\name{executeCohortCovariateCharacterization}
\alias{executeCohortCovariateCharacterization}
\title{Execute the cohort covariate}
\usage{
executeCohortCovariateCharacterization(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  targetCohortDatabaseSchema,
  targetCohortIds,
  covariateCohortDatabaseSchema,
  covariateCohortIds,
  unionCovariateCohorts = TRUE,
  covariateCohortDefinitionSet =
    dplyr::select(dplyr::filter(PhenotypeLibrary::getPhenotypeLog(), cohortId \%in\%
    c(covariateCohortIds)), cohortId, cohortName),
  targetCohortTableName,
  cohortCovariateAnalysisId = 150,
  covariateCohortTableName,
  temporalStartDays = c(-1:1),
  temporalEndDays = temporalStartDays,
  minCellCount = 5,
  minCharacterizationMean = 1e-04,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  databaseId = NULL,
  notes = NULL,
  outputFolder
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This
is commonly the same as cdmDatabaseSchema. Note that for
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{targetCohortDatabaseSchema}{Schema with instantiated target cohorts.}

\item{targetCohortIds}{Target Cohort ids.}

\item{covariateCohortDatabaseSchema}{Schema with instantiated target cohorts.}

\item{covariateCohortIds}{CohortIds of the covariate Cohorts. These must be instantiated
in covariateCohortDatabaseSchema.}

\item{unionCovariateCohorts}{Do you want to union the covariateCohorts into one cohort. If yes,
a new cohort will be created and that cohort will be called Composite
with the cohortId = -1.}

\item{covariateCohortDefinitionSet}{Cohort Definition set object of the covariate cohort}

\item{targetCohortTableName}{target Cohort Table Names}

\item{covariateCohortTableName}{Covariate Cohort Table Names}

\item{temporalStartDays}{A list of integers representing the start of a time period,
relative to the index date. 0 indicates the index date, -1
indicates the day before the index date, etc. The start day
is included in the time period.}

\item{temporalEndDays}{A list of integers representing the end of a time period,
relative to the index date. 0 indicates the index date, -1
indicates the day before the index date, etc. The end day
is included in the time period.}

\item{minCellCount}{Default 5}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not
truly support temp tables. To emulate temp tables,
provide a schema with write privileges where temp tables
can be created.}

\item{databaseId}{The id of the database}

\item{notes}{Do you want to add any text notes to metadata output}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}
}
\description{
Execute the cohort covariate
}
\details{
This function executes the cohort covariate
}
