% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteCohortFeatureExtraction.R
\name{executeCohortFeatureExtraction}
\alias{executeCohortFeatureExtraction}
\title{Execute feature extraction}
\usage{
executeCohortFeatureExtraction(
  connectionDetails,
  connection,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortIds,
  cohortTable,
  covariateCohortDatabaseSchema = cohortDatabaseSchema,
  covariateCohortIds = covariateCohortDefinitionSet$cohortId,
  covariateCohortTable = cohortTable,
  covariateCohortDefinitionSet,
  cohortCovariateAnalysisId = 150,
  temporalStartDays = c(-1:1),
  temporalEndDays = temporalStartDays,
  minCellCount = 5,
  aggregated = TRUE,
  minCharacterizationMean = 1e-04,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outputFolder = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}}

\item{connection}{An optional database connection object.
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema with instantiated target cohorts.}

\item{cohortIds}{cohort ids.}

\item{cohortTable}{cohort Table Names}

\item{aggregated}{Should aggregate statistics be computed instead of covariates per cohort entry?}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not
truly support temp tables. To emulate temp tables,
provide a schema with write privileges where temp tables
can be created.}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}

\item{covariateSettings}{FeatureExtraction covariateSettings object}
}
\description{
Execute feature extraction
}
\details{
This function executes the cohort covariate
}
