% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteCohortGeneration.R
\name{executeCohortGeneration}
\alias{executeCohortGeneration}
\title{Execute the cohort generation}
\usage{
executeCohortGeneration(
  connectionDetails,
  cohortDefinitionSet,
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTableNames = CohortGenerator::getCohortTableNames(cohortTable = "cohort"),
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  outputFolder,
  databaseId,
  createCohortTableIncremental = TRUE,
  generateCohortIncremental = TRUE,
  incrementalFolder = file.path(outputFolder, "incrementalFolder"),
  cohortIds = NULL
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cohortDefinitionSet}{Cohort Definition set object}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You
will need to have write privileges in this schema. Note
that for SQL Server, this should include both the
database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTableNames}{cohortTableNames}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not
truly support temp tables. To emulate temp tables,
provide a schema with write privileges where temp tables
can be created.}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}

\item{databaseId}{database id}

\item{createCohortTableIncremental}{When set to TRUE, this function will check to see if the
cohortTableNames exists in the cohortDatabaseSchema and if
they exist, it will skip creating the tables.}

\item{generateCohortIncremental}{Create only cohorts that haven't been created before?}

\item{incrementalFolder}{Name of local folder to hold the logs for incremental
run; make sure to use forward slashes (/). Do not use a
folder on a network drive since this greatly impacts
performance.}

\item{cohortIds}{Do you want to limit the execution to only some cohort ids.}
}
\description{
Execute the cohort generation
}
\details{
This function executes the cohort generation
}
