% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecuteConceptCovariateCharacterization.R
\name{executeConceptCovariateCharacterization}
\alias{executeConceptCovariateCharacterization}
\title{Execute the cohort covariate}
\usage{
executeConceptCovariateCharacterization(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  targetCohortDatabaseSchema,
  targetCohortIds,
  targetCohortTableName,
  temporalCovariateSettings = CohortDiagnostics:::getDefaultCovariateSettings(),
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  databaseId = NULL,
  minCellCount = 5,
  minCharacterizationMean = 1e-05,
  notes = NULL,
  outputFolder
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created
using the
\code{\link[DatabaseConnector]{createConnectionDetails}}
function in the DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM
format resides. Note that for SQL Server, this should
include both the database and schema name, for example
'cdm_data.dbo'.}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary data resides. This
is commonly the same as cdmDatabaseSchema. Note that for
SQL Server, this should include both the database and
schema name, for example 'vocabulary.dbo'.}

\item{targetCohortDatabaseSchema}{Schema with instantiated target cohorts.}

\item{targetCohortIds}{Target Cohort ids.}

\item{targetCohortTableName}{target Cohort Table Names}

\item{temporalCovariateSettings}{FeatureExtraction temporalCovariateSettings object}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not
truly support temp tables. To emulate temp tables,
provide a schema with write privileges where temp tables
can be created.}

\item{databaseId}{The id of the database}

\item{minCellCount}{Default 5}

\item{minCharacterizationMean}{Default 0.00001}

\item{notes}{Do you want to add any text notes to metadata output}

\item{outputFolder}{Name of local folder to place results; make sure to use
forward slashes (/). Do not use a folder on a network
drive since this greatly impacts performance.}
}
\description{
Execute the cohort covariate
}
\details{
This function executes the cohort covariate
}
