% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSectionsFromPmcXml.R
\name{extractSectionsFromPmcXml}
\alias{extractSectionsFromPmcXml}
\title{Extract Sections from PMC XML}
\usage{
extractSectionsFromPmcXml(pmcXml)
}
\arguments{
\item{pmcXml}{A character string containing the XML data or path to the XML file.}
}
\value{
A list containing the texts of the Abstract, Introduction, Methods, Results,
Discussion sections, and tables with their respective captions and data.
}
\description{
This function parses an XML document from PubMed Central (PMC) and extracts text content
from predefined sections such as Abstract, Introduction, Methods, Results, and Discussion.
It also handles extraction of tables with captions. The function is case-insensitive to
section titles.
}
