% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCirceHumanReadableCohortText.R
\name{getCirceHumanReadableCohortText}
\alias{getCirceHumanReadableCohortText}
\title{Returns list with circe generated cohort documentation}
\usage{
getCirceHumanReadableCohortText(
  cohortDefinition,
  cohortName = "Cohort Definition",
  includeConceptSets = FALSE,
  addUsefulHeaders = TRUE
)
}
\arguments{
\item{cohortDefinition}{An R object (list) with a list representation of the cohort definition expression,
that may be converted to a cohort expression JSON using
RJSONIO::toJSON(x = cohortDefinition, digits = 23, pretty = TRUE)}

\item{cohortName}{Name for the cohort definition}

\item{includeConceptSets}{Do you want to included concept set in the documentation}

\item{addUsefulHeaders}{Do you want useful headers like cohort name?}
}
\value{
list object
}
\description{
Returns list with circe generated cohort documentation
}
