% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCovariateSummaryByRowIdTimeId.R
\name{getCovariateSummaryByRowIdTimeId}
\alias{getCovariateSummaryByRowIdTimeId}
\title{Get temporal overlap between covariate and target cohorts}
\usage{
getCovariateSummaryByRowIdTimeId(
  covariateData,
  exitCohortIds = NULL,
  entryCohortIds = NULL,
  eventCohortIds
)
}
\arguments{
\item{covariateData}{A feature extraction output.}

\item{exitCohortIds}{Optional vector of exit cohort IDs.}

\item{entryCohortIds}{Optional vector of entry cohort IDs.}

\item{eventCohortIds}{vector of event cohort IDs.}
}
\value{
A data frame containing the covariate data.
}
\description{
This function assesses the temporal relationships between a target cohort
and multiple covariate cohorts by looking for overlap of covariate cohort
in a time window in relation to target cohort start date.
}
\details{
Overlap is calculated using FeatureExtraction.
}
\examples{
# Example usage
getCovariateSummaryByRowIdTimeId(connection = dbConnection,
                         connectionDetails = connectionDetails,
                         cohortDatabaseSchema = "myCohortSchema",
                         cdmDatabaseSchema = "myCdmSchema")
}
