% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFilesNamesThatMatchPattern.R
\name{getFilesNamesThatMatchPattern}
\alias{getFilesNamesThatMatchPattern}
\title{Retrieve file names that match any of the given patterns.}
\usage{
getFilesNamesThatMatchPattern(patterns, chunkSize = 1000)
}
\arguments{
\item{patterns}{A character vector containing file name patterns to match against.}

\item{chunkSize}{The number of patterns to process in each regex compilation, defaults to 1000.}
}
\value{
A character vector containing the full paths of files that match any of the provided patterns.
}
\description{
This function searches for files within the current working directory (and its subdirectories)
that match any of the patterns provided. Due to performance considerations with large sets of patterns,
the patterns are processed in chunks.
}
