% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGpt4Response.R
\name{getGpt4Response}
\alias{getGpt4Response}
\title{Communicate with a local instance of GPT-4}
\usage{
getGpt4Response(
  prompt,
  systemPrompt = "",
  additionalText = NULL,
  dataFrame = NULL,
  url = keyring::key_get("genai_gpt4_endpoint"),
  apiKey = keyring::key_get("genai_api_gpt4_key"),
  temperature = 1e-08,
  frequencyPenalty = 0,
  presencePenalty = 0,
  maxTokens = 4096
)
}
\arguments{
\item{prompt}{(required) A character string containing the user's prompt.}

\item{systemPrompt}{(optional) A character string containing the system prompt to set the context or behavior of the
assistant. Default is an empty string.}

\item{additionalText}{(optional) A blob of lot of text. e.g. read from one or more pdf files.}

\item{dataFrame}{(optional) A data frame to be converted to JSON and used as the prompt. Default is NULL.}

\item{url}{A character string containing the URL of the GPT-4 endpoint. Retrieved from keyring by default.}

\item{apiKey}{A character string containing the API key for the GPT-4 endpoint. Retrieved from keyring by default.}

\item{temperature}{(optional) A numeric value controlling the randomness of the response. Default is 0.00000001.}

\item{frequencyPenalty}{(optional) A numeric value affecting the likelihood of repeating the same lines or phrases. Default is 0.}

\item{presencePenalty}{(optional) A numeric value affecting the likelihood of introducing new topics. Default is 0.}

\item{maxTokens}{(optional) An integer specifying the maximum number of tokens to generate in the response. Default is 4096.}
}
\value{
A array of objects with output from GPT-4. Use "output" for the final text output.
}
\description{
This function sends a prompt to a local instance of GPT-4 and retrieves the response.
It supports input from direct text, a PDF file, or a data frame.
The function allows customization of various parameters for the GPT-4 API request.
}
