% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPubMedPmcIdFromPmId.R
\name{getPubMedPmcIdFromPmId}
\alias{getPubMedPmcIdFromPmId}
\title{Retrieve the PMCID for a given PubMed ID (PMID)}
\usage{
getPubMedPmcIdFromPmId(pmid)
}
\arguments{
\item{pmid}{A character string specifying the PubMed ID for which the PMCID is sought.}
}
\value{
A character string representing the PMCID associated with the given PMID.
}
\description{
This function queries the PubMed database using the NCBI's E-utilities API to fetch the corresponding
PMCID for a given PMID. The result is extracted from the XML response.
}
