% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPubMedPmcIdsFromPmIds.R
\name{getPubMedPmcIdsFromPmIds}
\alias{getPubMedPmcIdsFromPmIds}
\title{Retrieve PMCIDs for a vector of PubMed IDs (PMIDs)}
\usage{
getPubMedPmcIdsFromPmIds(pmids)
}
\arguments{
\item{pmids}{A character vector specifying the PubMed IDs for which PMCIDs are sought.}
}
\value{
A data frame with columns 'PMID' and 'PMCID', representing the PMCID associated with each given PMID.
}
\description{
This function queries the PubMed database using the NCBI's E-utilities API to fetch the corresponding
PMCIDs for given PMIDs. The results are extracted from the XML response and returned in a data frame.
}
