% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPubmedQuerySearchString.R
\name{getPubmQuerySearchString}
\alias{getPubmQuerySearchString}
\title{Get PubMed Query Search String}
\usage{
getPubmQuerySearchString(meshTerm, affiliatedJnJ = FALSE)
}
\arguments{
\item{meshTerm}{Character string specifying the main MeSH term to include in the search.}

\item{affiliatedJnJ}{Logical, if TRUE includes filters to only show articles affiliated with
Johnson & Johnson and excludes certain publication types like clinical trials.}
}
\value{
A character string representing the complete PubMed search query.
}
\description{
Constructs a search query string for PubMed based on specified search criteria. This function creates
a default search string using various epidemiological and cohort study related terms. It can be extended
to include specific affiliation filters, such as for Johnson & Johnson.
}
\examples{
getPubmQuerySearchString("diabetes", TRUE)
}
