% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/GetTargetCohortToFeatureCohortTemporalRelationship.R
\name{getTargetCohortToFeatureCohortTemporalRelationship}
\alias{getTargetCohortToFeatureCohortTemporalRelationship}
\title{Function to calculate cohort periods}
\usage{
getTargetCohortToFeatureCohortTemporalRelationship(
  connectionDetails = NULL,
  connection = NULL,
  targetCohortDatabaseSchema,
  featureCohortDatabaseSchema = targetCohortDatabaseSchema,
  targetCohortTableName,
  featureCohortTableName = targetCohortTableName,
  targetCohortId,
  featureCohortIds,
  returnCohortData = TRUE
)
}
\arguments{
\item{connectionDetails}{Connection details, NULL by default}

\item{connection}{Database connection, NULL by default}

\item{targetCohortDatabaseSchema}{Database schema for target cohort}

\item{featureCohortDatabaseSchema}{Database schema for feature cohort, default to targetCohortDatabaseSchema}

\item{targetCohortTableName}{Table name for target cohort}

\item{featureCohortTableName}{Table name for feature cohort, default to targetCohortTableName}

\item{targetCohortId}{ID for target cohort}

\item{featureCohortIds}{IDs for feature cohorts}

\item{returnCohortData}{Boolean to decide if cohort data should be returned, TRUE by default}
}
\value{
A list of cohort data and relationships
}
\description{
Function to calculate cohort periods
}
