% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParsePubMedXML.R
\name{parsePubMedXML}
\alias{parsePubMedXML}
\title{Parse XML Data from PubMed}
\usage{
parsePubMedXML(xml_data)
}
\arguments{
\item{xml_data}{XML document containing PubMed article data.}
}
\value{
A tibble with columns for Title, Publication Date, Authors, and PMID of each article.
}
\description{
This function parses XML data obtained from PubMed to extract information
about articles. The expected XML format should include specific nodes such as
PubmedArticle, ArticleTitle, PubDate, Author, LastName, Initials, and PMID.
}
