% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareEventDistributionDataForViolinPlots.R
\name{prepareEventDistributionDataForViolinPlots}
\alias{prepareEventDistributionDataForViolinPlots}
\title{Prepare Data for Violin Plots}
\usage{
prepareEventDistributionDataForViolinPlots(
  data,
  startSequence = NULL,
  minDaysToNextCohortStart = NULL,
  maxDaysToNextCohortStart = NULL
)
}
\arguments{
\item{data}{A data frame containing the columns \code{startSequence}, \code{daysToNextCohortStart}, \code{subjects}, \code{databaseId}, and \code{subjectsTotal}. This is the dataset to be prepared for the violin plot.}

\item{startSequence}{An integer indicating the start sequence to filter the data by. If \code{-1}, no filtering on start sequence will be applied. Default is \code{-1}.}

\item{minDaysToNextCohortStart}{An optional integer specifying the minimum number of days to the next cohort start for filtering the data. Rows with \code{daysToNextCohortStart} less than this value will be excluded. If \code{NULL}, no minimum limit will be applied.}

\item{maxDaysToNextCohortStart}{An optional integer specifying the maximum number of days to the next cohort start for filtering the data. Rows with \code{daysToNextCohortStart} greater than this value will be excluded. If \code{NULL}, no maximum limit will be applied.}
}
\value{
A data frame prepared for violin plot visualization with columns \code{value} (formerly \code{daysToNextCohortStart}), \code{name} (formerly \code{databaseId}), and \code{myaxis} (a concatenation of \code{databaseId} and \code{subjectsTotal} for labeling).
}
\description{
This function prepares event distribution data specifically for violin plot visualization. It filters and transforms the data based on the provided criteria related to cohort start sequences and the number of days to the next cohort start.
}
\examples{
preparedData <- prepareEventDistributionDataForViolinPlots(myData, -1, NULL, NULL)
}
