% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedQuery.R
\name{pubMedQuery}
\alias{pubMedQuery}
\title{Query PubMed Database}
\usage{
pubMedQuery(searchText, affiliatedJnJ = FALSE)
}
\arguments{
\item{searchText}{Character string, the main term to include in the search.}

\item{affiliatedJnJ}{Logical, set to TRUE to include Johnson & Johnson affiliated articles,
filtering out certain publication types (e.g., clinical trials).}
}
\value{
An XML document.
}
\description{
This function queries the PubMed database using specified search criteria. It supports
a default search on various epidemiological and cohort terms and can be extended
to include specific affiliation filters such as Johnson & Johnson.
}
