% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrieveAndSavePmcXmlForPmcIds.R
\name{retrieveAndSavePmcXmlForPmcIds}
\alias{retrieveAndSavePmcXmlForPmcIds}
\title{Retrieve and Save PMC XML Using Multiple Methods}
\usage{
retrieveAndSavePmcXmlForPmcIds(pmcIds, outputFolder = ".", incremental = TRUE)
}
\arguments{
\item{pmcIds}{A vector of PMC IDs for which to retrieve XML data.}

\item{outputFolder}{A character string specifying the output folder location. The default is
the current working directory.}

\item{incremental}{A logical value indicating whether to skip retrieval if the output file already
exists. Default is \code{TRUE}.}
}
\value{
A list with two elements:
\itemize{
\item{success}{A list of successfully retrieved and saved XML contents, indexed by PMC ID.}
\item{failed}{A list of PMC IDs for which XML retrieval failed using both methods, set to \code{NA}.}
}
}
\description{
This function attempts to retrieve XML data for a list of PMC IDs using the tidy PMC method first.
If the tidy PMC method fails, it then tries using the Rentrez method. It saves the retrieved XML
to files in directories corresponding to the retrieval method used within a specified output folder.
}
