% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrievePMCXML.R
\name{retrievePMCXML}
\alias{retrievePMCXML}
\title{Retrieve PMC XML}
\usage{
retrievePMCXML(pmcId)
}
\arguments{
\item{pmcId}{A character string representing the PMC ID of the article.}
}
\value{
A character string containing the full text of the article in XML format.
}
\description{
This function fetches the full text of an article in XML format from PubMed Central (PMC) using a given PMC ID.
}
