% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrievePMCXMLUsingRentrez.R
\name{retrievePMCXMLUsingRentrez}
\alias{retrievePMCXMLUsingRentrez}
\title{Retrieve PMC XML}
\usage{
retrievePMCXMLUsingRentrez(pmcId)
}
\arguments{
\item{pmcId}{A character string representing the PMC ID of the article.}
}
\value{
A character string containing the full text of the article in XML format, or an error message if unable to retrieve.
}
\description{
This function fetches the full text of an article in XML format from PubMed Central (PMC) using a given PMC ID.
This function employs the rentrez package to fetch the XML. If an error occurs during the fetching process,
an appropriate error message is returned.
}
