% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RetrievePMCXMLUsingTidyPmc.R
\name{retrievePMCXMLUsingTidyPmc}
\alias{retrievePMCXMLUsingTidyPmc}
\title{Retrieve PMC XML}
\usage{
retrievePMCXMLUsingTidyPmc(pmcId)
}
\arguments{
\item{pmcId}{A character string representing the PMC ID of the article.}
}
\value{
A character string containing the full text of the article in XML format, or an error message if unable to retrieve.
}
\description{
This function fetches the full text of an article in XML format from PubMed Central (PMC) using a given PMC ID.
If the direct retrieval fails, it searches for an open access article via europepmc and attempts to download it.
}
