% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDataFrameToExcel.R
\name{saveDataFrameToExcel}
\alias{saveDataFrameToExcel}
\title{Save a DataFrame to an Excel File with Formatting}
\usage{
saveDataFrameToExcel(dataFrame, filePath, sheetName = "Sheet 1")
}
\arguments{
\item{dataFrame}{The dataframe to be saved to Excel.}

\item{filePath}{The file path where the Excel file will be saved.}

\item{sheetName}{The name of the worksheet in the Excel file (default is "Sheet 1").}
}
\description{
This function saves a given dataframe to an Excel file. It adjusts the width of the columns.
}
