% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TracePaths.R
\name{tracePath}
\alias{tracePath}
\title{Helper function to trace paths for a given root using DFS}
\usage{
tracePath(df, root)
}
\arguments{
\item{df}{A dataframe representing edges with 'parentId' and 'childId' columns.}

\item{root}{The root node from which to start tracing paths.}
}
\value{
A data frame representing all paths originating from the root.
}
\description{
This function constructs paths for a specified root by performing a DFS on a graph represented as a dataframe.
}
