% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMultipleCohortSubsetDefinitions.R
\name{addMultipleCohortSubsetDefinitions}
\alias{addMultipleCohortSubsetDefinitions}
\title{Add multiple cohort subset definitions}
\usage{
addMultipleCohortSubsetDefinitions(
  cohortDefinitionSet,
  targetCohortIds,
  namePrefix,
  ...
)
}
\arguments{
\item{cohortDefinitionSet}{A cohort definition set to which the subset definitions will be added.}

\item{targetCohortIds}{A vector of target cohort IDs.}

\item{namePrefix}{A prefix for the name of each cohort subset definition.}

\item{...}{One or more vectors of subset operators.}
}
\value{
The modified cohort definition set with added subset definitions.
}
\description{
This function adds multiple cohort subset definitions to a given cohort definition set.
It iterates over provided sequences of subset operators in a nested manner,
creating a unique subset definition for every possible combination of operators.
}
\examples{
cohortDefinitionSetWithSubset <- addMultipleCohortSubsetDefinitions(
  cohortDefinitionSet,
  targetOrIndicationCohortIds,
  subsetOperatorSequence1,
  subsetOperatorSequence2
  # ... Add more sequences if necessary
)
}
