% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCohortTemporalOverlap.R
\name{getCohortTemporalOverlap}
\alias{getCohortTemporalOverlap}
\title{Get temporal overlap between covariate and target cohorts}
\usage{
getCohortTemporalOverlap(
  connection = NULL,
  connectionDetails = NULL,
  cohortDatabaseSchema,
  cdmDatabaseSchema,
  cohortCovariateAnalysisId = 150,
  cohortTable,
  covariateCohortDefinitionSet,
  covariateCohortIds,
  exitCohortIds = NULL,
  entryCohortIds = NULL,
  startDays,
  endDays,
  targetCohortId
)
}
\arguments{
\item{connection}{An optional database connection object.}

\item{connectionDetails}{An object containing the database connection details.}

\item{cohortDatabaseSchema}{The name of the database schema containing the cohort.}

\item{cdmDatabaseSchema}{The name of the CDM database schema.}

\item{cohortCovariateAnalysisId}{The ID for cohort covariate analysis (default 150).}

\item{cohortTable}{The name of the table containing the cohort data.}

\item{covariateCohortDefinitionSet}{Definitions for covariate cohorts.}

\item{covariateCohortIds}{A vector of IDs for the covariate cohorts.}

\item{exitCohortIds}{Optional vector of exit cohort IDs.}

\item{entryCohortIds}{Optional vector of entry cohort IDs.}

\item{targetCohortId}{The ID of the target cohort.}
}
\value{
A data frame containing the covariate data.
}
\description{
This function assesses the temporal relationships between a target cohort
and multiple covariate cohorts by looking for overlap of covariate cohort
in a time window in relation to target cohort start date.
}
\details{
Overlap is calculated using FeatureExtraction.
}
\examples{
# Example usage
getCohortTemporalOverlap(connection = dbConnection,
                         connectionDetails = connectionDetails,
                         cohortDatabaseSchema = "myCohortSchema",
                         cdmDatabaseSchema = "myCdmSchema",
                         targetCohortId = 123)
}
